fix: change grid type int to double

+++ ouds_theme_contract/lib/theme/scheme/responsive/ouds_grid_scheme.dart
@@ -59,7 +59,7 @@ class OudsGridScheme {
         medium: gridTokens.mediumMargin,
       );
 
-  int get columnGap => OudsWindowSizeClassUtil.select(
+  double get columnGap => OudsWindowSizeClassUtil.select(
         sizeClass: sizeClass,
         extraCompact: gridTokens.extraCompactColumnGap,
         compact: gridTokens.compactColumnGap,
